components

+++ components/brave_ads/core/internal/account/utility/refill_confirmation_tokens/refill_confirmation_tokens_util.cc
@@ -14,7 +14,7 @@ bool ShouldRefillConfirmationTokens() {
   return ConfirmationTokenCount() < kMinConfirmationTokens.Get();
 }
 
-int CalculateAmountOfConfirmationTokensToRefill() {
+size_t CalculateAmountOfConfirmationTokensToRefill() {
   return kMaxConfirmationTokens.Get() - ConfirmationTokenCount();
 }
